docker_compose/movies/app/requirements.txt
django==4.2.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-split-settings==1.2.0
django_debug_toolbar==4.2.0
uWSGI==2.0.23
